Fixed function declaration
git-svn-id: https://parflow.svn.cloudforge.com/parflow/parflow/trunk@423 39af3fb0-ca4e-0410-adf9-b4ae76d04516

+++ pfsimulator/clm/drv_gridave.F90
@@ -56,7 +56,7 @@ function drv_gridave (nch, mask, fgr, var, drv)
 
 end function drv_gridave
 
-real(r8) function variance(nch,var)
+real*8 function variance(nch,var)
   !=========================================================================
   !
   !  CLMCLMCLMCLMCLMCLMCLMCLMCL  A community developed and sponsored, freely
